+++ service/album_service.go
@@ -15,10 +15,10 @@ func NewAlbumService(repo *repository.AlbumRepository) *AlbumService {
 	return &AlbumService{repo: repo}
 }
 
-func (s *AlbumService) CreateAlbum(ctx context.Context, album models.Album) (int, error) {
+func (s *AlbumService) CreateAlbum(ctx context.Context, album models.Album) (uint, error) {
 	return s.repo.CreateAlbum(ctx, album)
 }
 
-func (s *AlbumService) GetAlbumByID(ctx context.Context, id int64) (*models.Album, error) {
+func (s *AlbumService) GetAlbumByID(ctx context.Context, id uint) (*models.Album, error) {
 	return s.repo.GetAlbumByID(ctx, id)
 }
